# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/people/{uuidP}/{uuidT}/complete": {
            "get": {
                "description": "Завершает задачу для человека по его UUID и UUID задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Завершение задачи для человека",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID человека",
                        "name": "uuidP",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID задачи",
                        "name": "uuidT",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное завершение задачи",
                        "schema": {
                            "$ref": "#/definitions/models.Ok"
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Человек или задача с указанным UUID не найдены",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/people/{uuidP}/{uuidT}/pause": {
            "get": {
                "description": "Приостанавливает таймер для задачи указанного человека по его UUID и UUID задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Приостановка таймера для задачи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID человека",
                        "name": "uuidP",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID задачи",
                        "name": "uuidT",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная приостановка таймера для задачи",
                        "schema": {
                            "$ref": "#/definitions/models.Ok"
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Человек или задача с указанным UUID не найдены",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/people/{uuidP}/{uuidT}/start": {
            "get": {
                "description": "Начинает таймер для задачи указанного человека по его UUID и UUID задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Начало таймера для задачи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID человека",
                        "name": "uuidP",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID задачи",
                        "name": "uuidT",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное начало таймера для задачи",
                        "schema": {
                            "$ref": "#/definitions/models.Ok"
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Человек или задача с указанным UUID не найдены",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/people/{uuid}/create-task": {
            "post": {
                "description": "Создает новую задачу для человека по его UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создание новой задачи для человека",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID человека",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для создания задачи",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TaskCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданная задача",
                        "schema": {
                            "$ref": "#/definitions/models.Ok"
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Человек с указанным UUID не найден",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/people/{uuid}/worktime": {
            "post": {
                "description": "Возвращает список задач с временем работы в указанном диапазоне для указанного человека по его UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получение списка задач с временем работы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID человека",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Временной диапазон",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DateStartEnd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач с временем работы",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TaskTimeResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Человек с указанным UUID не найден или нет задач в указанном диапазоне",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.TaskTimeResult": {
            "type": "object",
            "properties": {
                "idtask": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                },
                "total_time": {
                    "type": "string"
                }
            }
        },
        "models.DateStartEnd": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Ok": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "models.TaskCreate": {
            "type": "object",
            "properties": {
                "idPerson": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                }
            }
        }
    }
}